+++ lib/pact_broker/client/base_command.rb
@@ -90,8 +90,20 @@ module PactBroker
         ::Term::ANSIColor.green(text)
       end
 
+      def blue(text)
+        ::Term::ANSIColor.blue(text)
+      end
+
+      def magenta(text)
+        ::Term::ANSIColor.magenta(text)
+      end
+
       def red(text)
         ::Term::ANSIColor.red(text)
+      end
+
+      def yellow(text)
+        ::Term::ANSIColor.yellow(text)
       end
     end
   end