+++ spec/service_providers/pactflow_publish_provider_contract_spec.rb
@@ -2,7 +2,7 @@ require_relative "pact_helper"
 require "pactflow/client/provider_contracts/publish"
 require "yaml"
 
-RSpec.describe "publishing a provider contract to Pactflow", pact: true do
+RSpec.describe "publishing a provider contract to PactFlow", pact: true do
   before do
     # no point re-testing this
     allow(PactBroker::Client::Versions::Create).to receive(:call).and_return(double("result", success: true))
@@ -87,6 +87,41 @@ RSpec.describe "publishing a provider contract to Pactflow", pact: true do
       expect(subject).to be_a PactBroker::Client::CommandResult
       expect(subject.success).to be true
       expect(subject.message).to include "Successfully published provider contract for Bar version 1"
+      expect(subject.message).not_to include pactflow.mock_service_base_url
+    end
+  end
+
+  context "creating a provider contract with valid parameters with pf:ui return results" do
+    let(:success_response_with_pf_ui_url) do
+      {
+        status: response_status,
+        headers: pact_broker_response_headers,
+        body: { "_links": {
+          "pf:ui": {
+            "href": "#{pactflow.mock_service_base_url}/contracts/bi-directional/provider/Bar/version/1/provider-contract"
+          }
+        } }
+      }
+    end
+    before do
+      pactflow
+        .given("there is a pf:ui href in the response")
+        .upon_receiving("a request to create a provider contract")
+        .with(
+          method: :put,
+          path: "/contracts/provider/Bar/version/1",
+          headers: put_request_headers,
+          body: request_body
+        )
+        .will_respond_with(success_response_with_pf_ui_url)
+    end
+
+    it "returns a CommandResult with success = true and a provider contract ui url" do
+      expect(subject).to be_a PactBroker::Client::CommandResult
+      expect(subject.success).to be true
+      expect(subject.message).to include "Successfully published provider contract for Bar version 1"
+      expect(subject.message).to include "Next steps:"
+      expect(subject.message).to include success_response_with_pf_ui_url[:body][:_links][:'pf:ui'][:href]
     end
   end
 end